+++ src/pushsource/_impl/backend/staged/staged_ami.py
@@ -1,7 +1,7 @@
 import os
 import logging
 
-from ...model import AmiRelease, AmiPushItem
+from ...model import AmiRelease, AmiPushItem, AmiBillingCodes
 from .staged_base import StagedBaseMixin, handles_type
 
 LOG = logging.getLogger("pushsource")
@@ -33,6 +33,20 @@ class StagedAmiMixin(StagedBaseMixin):
 
         release = AmiRelease(**release_kwargs)
 
+        image_kwargs = {"release": release}
+
+        billing_codes_md = attributes.get("billing_codes") or {}
+        if billing_codes_md:
+            billing_codes_attrs = ["name", "codes"]
+            billing_codes_kwargs = {}
+
+            for key in billing_codes_attrs:
+                if key in billing_codes_md:
+                    billing_codes_kwargs[key] = billing_codes_md[key]
+
+            billing_codes = AmiBillingCodes(**billing_codes_kwargs)
+            image_kwargs.update({"billing_codes": billing_codes})
+
         image_attrs = [
             "type",
             "region",
@@ -43,7 +57,7 @@ class StagedAmiMixin(StagedBaseMixin):
             "sriov_net_support",
             "ena_support",
         ]
-        image_kwargs = {"release": release}
+
         for key in image_attrs:
             if key in attributes:
                 image_kwargs[key] = attributes[key]