+++ src/pushsource/_impl/model/__init__.py
@@ -11,4 +11,4 @@ from .file import FilePushItem
 from .modulemd import ModuleMdPushItem
 from .comps import CompsXmlPushItem
 from .productid import ProductIdPushItem
-from .ami import AmiPushItem, AmiRelease
+from .ami import AmiPushItem, AmiRelease, AmiBillingCodes